+++ Keg/src/Renderer/Buffers/OpenGLVAO.h
@@ -10,7 +10,7 @@ namespace Keg
 
 		virtual void Bind();
 		virtual void UnBind();
-		virtual void LinkAttrib(uint32_t layout, uint32_t size, uint32_t type, int stride, uint32_t offset);
+		virtual void LinkAttrib(uint32_t layout, uint32_t size, uint32_t type, int stride, uint64_t offset);
 		inline uint32_t GetID() { return m_ID; }
 
 	private:
